requirements.txt
kafka-python
Flask
pymongo
folium
python-dotenv
pybloom_live